requirements.txt
dj-database-url==0.4.
django-cors-headers==1.1.0
opbeat==3.3.4
gunicorn==19.4.5
whitenoise==2.0.6
Django==1.9
django-allauth==0.26.1
django-rest-auth==0.8.1
django-extra-fields==0.7
djangorestframework==3.4.3
oauthlib==1.1.2
psycopg2==2.6.2
python-openid==2.2.5
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
test==2.3.4.5
Pillow==3.3.1
django-filter==0.14.0
djangorestframework-filters==0.8.0
python-social-auth==0.2.21
rest-social-auth==0.5.0
djangorestframework-jwt==1.8.0















